Move TemplateWidgetConfigureUiState and TemplateRenderError to separate files
* Extract TemplateWidgetConfigureUiState and TemplateRenderError to a new file
* Handle CancellationException during template rendering in TemplateWidgetConfigureViewModel
